requirements.txt
fastapi~=0.128.0
uvicorn[standard]~=0.40.0
pydantic~=2.12.5
sqlalchemy>=1.4
psycopg2-binary
redis~=7.1.0
httpx
celery~=5.6.2
python-dotenv~=1.2.1
requests~=2.32.3
tenacity~=9.1.2